Add permission checking for sport activities

+++ backend/sport_hours/views/sport_activities.py
@@ -1,5 +1,6 @@
 from flask import request, abort
 from flask.views import MethodView
+from flask_login import login_required, current_user
 
 from sport_hours.extensions import db
 from sport_hours.blueprints import api
@@ -8,6 +9,7 @@ from sport_hours.schemas import SportActivitySchema, ClubSchema, UserSchema
 
 
 @api.route('/activities')
+@login_required
 def get_assigned_activities():
     student = User.query.get_or_404(request.args.get('assigned_to'))
     out_schema = SportActivitySchema(many=True, exclude=('assigned_students',))
@@ -15,6 +17,7 @@ def get_assigned_activities():
 
 
 @api.route('/activities/all')
+@login_required
 def get_all_activities():
     activities = SportActivity.query.all()
     out_schema = SportActivitySchema(many=True, exclude=('assigned_students',))
@@ -22,6 +25,7 @@ def get_all_activities():
 
 
 @api.route('/activities/clubs/all')
+@login_required
 def get_all_clubs():
     clubs = Club.query.all()
     out_schema = ClubSchema(many=True)
@@ -29,6 +33,7 @@ def get_all_clubs():
 
 
 @api.route('/activities/<int:activity_id>', methods=['GET'])
+@login_required
 def get_activity_by_id(activity_id):
     activity = SportActivity.query.get_or_404(activity_id)
 
@@ -37,13 +42,18 @@ def get_activity_by_id(activity_id):
 
 
 class AssignedStudentsAPI(MethodView):
+    @login_required
     def get(self, activity_id):
         '''Get a list of students assigned to an activity.'''
         activity = SportActivity.query.get_or_404(activity_id)
         out_schema = UserSchema(many=True, exclude=('roles', 'activities'))
         return out_schema.jsonify(activity.assigned_students)
 
+    @login_required
     def post(self, activity_id):
+        if not current_user.is_admin:
+            abort(401)
+        
         '''Assign a given student to a sport activity.'''
         if not request.is_json:
             abort(400, 'The request must be in JSON.')
@@ -58,7 +68,11 @@ class AssignedStudentsAPI(MethodView):
         db.session.commit()
         return ('', 204)
 
+    @login_required
     def delete(self, activity_id):
+        if not current_user.is_admin:
+            abort(401)
+        
         '''Unassign a given student from a sport activity.'''
         if not request.is_json:
             abort(400, 'The request must be in JSON.')
@@ -85,7 +99,11 @@ api.add_url_rule('/activities/<int:activity_id>/assigned',
 
 
 @api.route('/activities/create', methods=['POST'])
+@login_required
 def create_sport_activity():
+    if not current_user.is_admin():
+        abort(401)
+    
     in_schema = SportActivitySchema(exclude=('id',))
     activity_record = in_schema.load(request.json)
 
@@ -96,7 +114,11 @@ def create_sport_activity():
 
 
 @api.route('/activities/<int:activity_id>', methods=['PUT'])
+@login_required
 def modify_sport_activity(activity_id):
+    if not current_user.is_admin:
+        abort(401)
+    
     in_schema = SportActivitySchema(exclude=('id',))
     record_in_db = SportActivity.query.get_or_404(activity_id)
 
@@ -108,7 +130,11 @@ def modify_sport_activity(activity_id):
 
 
 @api.route('/activities/<int:activity_id>', methods=['DELETE'])
+@login_required
 def delete_activity(activity_id):
+    if not current_user.is_admin:
+        abort(401)
+    
     SportActivity.query.filter_by(id=activity_id).delete()
     db.session.commit()
 
@@ -116,7 +142,11 @@ def delete_activity(activity_id):
 
 
 @api.route('/activities/clubs/<int:club_id>', methods=['PUT'])
+@login_required
 def create_club(club_id):
+    if not current_user.is_admin:
+        abort(401)
+    
     SportActivity.query.get_or_404(club_id)
 
     in_schema = ClubSchema(exclude=('id',))
@@ -136,7 +166,11 @@ def create_club(club_id):
 
 
 @api.route('/activities/clubs/<int:club_id>', methods=['DELETE'])
+@login_required
 def delete_club(club_id):
+    if not current_user.is_admin:
+        abort(401)
+    
     Club.query.filter_by(id=club_id).delete()
     db.session.commit()
